Restructures serializers
