+++ keystone_api/apps/allocations/serializers/nested.py
@@ -1,0 +1,30 @@
+
+
+from rest_framework import serializers
+
+from ..models import *
+
+__all__ = [
+    'AllocationRequestSummarySerializer',
+    'ClusterSummarySerializer',
+]
+
+
+class ClusterSummarySerializer(serializers.ModelSerializer):
+    """Serializer for summarizing cluster names in nested responses."""
+
+    class Meta:
+        """Serializer settings."""
+
+        model = Cluster
+        fields = ['name', 'enabled']
+
+
+class AllocationRequestSummarySerializer(serializers.ModelSerializer):
+    """Serializer for summarizing allocation requests in nested responses."""
+
+    class Meta:
+        """Serializer settings."""
+
+        model = AllocationRequest
+        fields = ['title', 'status', 'active', 'expire']
