+++ keystone_api/apps/users/serializers/common.py
@@ -1,73 +1,17 @@
-"""Serializers for casting database models to/from JSON and XML representations.
-
-Serializers handle the casting of database models to/from HTTP compatible
-representations in a manner that is suitable for use by RESTful endpoints.
-They encapsulate object serialization, data validation, and database object
-creation.
-"""
-
 from django.contrib.auth import password_validation
 from django.contrib.auth.hashers import make_password
 from django.db import transaction
 from rest_framework import serializers
 
-from plugins.auditlog import AuditlogFieldSerializer
-from .models import *
+from ..models import *
+from .nested import *
 
 __all__ = [
     'MembershipSerializer',
     'PrivilegedUserSerializer',
     'RestrictedUserSerializer',
-    'TeamRoleSerializer',
     'TeamSerializer',
-    'TeamSummarySerializer',
-    'UserRoleSerializer',
-    'UserSummarySerializer',
 ]
-
-
-class UserSummarySerializer(serializers.ModelSerializer):
-    """Serializer for summarizing user information in nested responses."""
-
-    class Meta:
-        """Serializer settings."""
-
-        model = User
-        fields = ["username", "first_name", "last_name", "email"]
-
-
-class UserRoleSerializer(serializers.ModelSerializer):
-    """Serializer for summarizing team member usernames and roles in nested responses."""
-
-    _user = UserSummarySerializer(source="user", read_only=True)
-
-    class Meta:
-        """Serializer settings."""
-
-        model = Membership
-        fields = ["id", "user", "role", "_user"]
-
-
-class TeamSummarySerializer(serializers.ModelSerializer):
-    """Serializer for summarizing team information in nested responses."""
-
-    class Meta:
-        """Serializer settings."""
-
-        model = Team
-        fields = ["name", "is_active"]
-
-
-class TeamRoleSerializer(serializers.ModelSerializer):
-    """Serializer for summarizing team names and roles in nested responses."""
-
-    _team = TeamSummarySerializer(source="team", read_only=True)
-
-    class Meta:
-        """Serializer settings."""
-
-        model = Membership
-        fields = ["id", "team", "role", "_team"]
 
 
 class MembershipSerializer(serializers.ModelSerializer):
